requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
blinker==1.6.3
mysql-connector-python==8.2.0
bcrypt==4.1.1
Pillow==10.1.0
pyzbar==0.1.9
flask-mail==0.9.1
requests==2.31.0
python-dotenv==1.0.0
flask-compress==1.14